Adjacent test calculator refactored to use TileDirection enum, switch case removed
